Made crud in the NoteController class

+++ src/main/java/com/notepad/app/services/NoteService.java
@@ -3,7 +3,6 @@ package com.notepad.app.services;
 import com.notepad.app.exceptions.NoteNotFoundException;
 import com.notepad.app.models.Note;
 import com.notepad.app.repositories.NoteRepository;
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,13 +10,10 @@ import java.util.List;
 
 @Service
 public class NoteService {
-    private final ModelMapper mapper;
     private final NoteRepository noteRepository;
 
     @Autowired
-    public NoteService(ModelMapper mapper,
-                       NoteRepository noteRepository) {
-        this.mapper = mapper;
+    public NoteService(NoteRepository noteRepository) {
         this.noteRepository = noteRepository;
     }
 
@@ -25,34 +21,32 @@ public class NoteService {
         return noteRepository.save(note);
     }
 
-//    public List<Note> findAllNotesByTextContainingInTitleOrContent(String text) {
-//        return noteRepository.findAllByTextContainingInTitleOrContent(text);
-//    }
+    public List<Note> findAllNotesByText(String text) {
+        return noteRepository.findAllByTextInTitleOrContent(text);
+    }
 
     public Note findNoteById(Long id) {
-        String message = String.format("Note with id %d not found", id);
         return noteRepository
                 .findById(id)
-                .orElseThrow(() -> new NoteNotFoundException(message));
+                .orElseThrow(() -> {
+                    String message = getMsgNoteNotFoundBy(id);
+                    return new NoteNotFoundException(message);
+                });
     }
 
     public List<Note> findAllNotes() {
         return noteRepository.findAll();
     }
 
-//    public Note findNoteByIdAndUser(Long id, User user) {
-//        String msg = String.format("Note with id %d and user id %d not found",
-//                id, user.getId());
-//        return noteRepository
-//                .findByIdAndUser(id, user)
-//                .orElseThrow(() -> new NoteNotFoundException(msg));
-//    }
-
-//    public List<Note> findAllNotesByUser(User user) {
-//        return noteRepository.findAllByUser(user);
-//    }
-
     public void deleteNoteById(Long id) {
-        noteRepository.deleteById(id);
+        if (noteRepository.existsById(id)) {
+            noteRepository.deleteById(id);
+        } else {
+            throw new NoteNotFoundException(getMsgNoteNotFoundBy(id));
+        }
+    }
+
+    private String getMsgNoteNotFoundBy(Long id) {
+        return String.format("Note with id %d not found", id);
     }
 }
